controller and book layers were implemented added global exception handler

+++ src/main/java/com/anima/graphql/repository/BookRepository.java
@@ -3,5 +3,6 @@ package com.anima.graphql.repository;
 import com.anima.graphql.domain.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book, Long> {
+
 }
